Teleportation + Misc
Added basic teleportation stuff. Added a line when you teleport, also a raycast that does nothing at the moment. We can probably add the boss and then use the raycast for damaging it. I also made the teleporter sprites, not that good but we can maybe stick to a green dark steampunk-ish theme.